Checkboxes bugfix.
Removed element #title_display assignment of ‘input’.

According to  Drupal\Core\Render\Element\FormElement, input is no longer an option.

Existing options are: before, after, invisible, attribute.

The first three options would seem not to be options for the usecase and attribute behaved the same as input.

Removed pointer-events as this stopped any interaction with the checkbox elements.