Exo modal fixed float element interaction.
Currently the js is targeting fixed elements that are an active float. If a modal is triggered, this float is animated closed before opening the modal. Perfect.

If a fixed element is not floated, before interaction- this element will not be animated closed before the modal is opened, potentially causing overlap issues.

Specifically an issue with navigations. Generally these are highest on the z-index and may be unable to change due to content and menu hover visibility issues.

This adjustment will target the unfloated and floated element.