Caching adjustment.
Caching now is applied by default, if not an admin route, by context of users with view permission, and by query details of q, sort and order.

Setting removed for turning on caching, since its now automatically decided.

Cache options moved to bottom of settings and wrapped in a fieldset for organization.

New cache option added to cache with url context. If a display depends on data from the url, such as node id, caching mentioned above is not granular enough and will cause the same display for all nodes depending on which node is loaded first.